# public/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Space-Stylish E-commerce API
  description: |
    Space-Stylish E-commerce Website API Documentation

    This is a complete e-commerce platform API that provides the following features:
    - User registration, login and profile management
    - Product browsing, search and detailed information
    - Shopping cart and order management
    - Marketing campaign management
    - Sales reports and analytics
    - Admin functions

    ## Authentication
    Most APIs require JWT token authentication. Please include in the Authorization header:
    `Authorization: Bearer <your_jwt_token>`

    ## Rate Limiting
    - General APIs: 60 requests per second
    - API endpoints: 30 requests per second
    - Admin endpoints: 10 requests per second
    - Health check and static resources are excluded

    ## Response Format
    All APIs return a standardized response format:
    ```json
    {
      "success": true,
      "data": {...},
      "message": "Success message",
      "timestamp": "2024-01-01T00:00:00.000Z"
    }
    ```

    ## Environment Variables
    - `API_VERSION`: API version (default: v1)
    - `SERVER_PORT`: Server port
    - `JWT_SIGN_SECRET`: JWT signing secret
    - `PARTNER_KEY`: TapPay partner key
    - `MERCHANT_ID`: TapPay merchant ID
    - `APP_ID`: Facebook App ID
    - `APP_SECRET`: Facebook App Secret

  version: 2.0.0
  contact:
    name: Space-Stylish Team
    email: support@space-stylish.com
  license:
    name: MIT
...
servers:
...
paths:
...
  /api/v1/user/signup:
    post:
      tags:
        - Authentication
      summary: User registration
      description: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                  description: User name
                  example: "John Doe"
                email:
                  type: string
                  format: email
                  description: User email
                  example: "john@example.com"
                password:
                  type: string
                  description: User password
                  example: "password123"
      responses:
        "200":
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      access_token:
                        type: string
                        description: JWT access token
                      access_expired:
                        type: integer
                        description: Token expiration time (seconds)
                      user:
                        $ref: '#/components/schemas/User'
        "400":
          description: Request parameter error
          content:
            application/json:
              schema:
                type: string
                example: "email format is wrong"
        "403":
          description: Email already exists
          content:
            application/json:
              schema:
                type: string
                example: "email already exist"
  /api/v1/user/signin:
    post:
      tags:
        - Authentication
      summary: User login
      description: User login (supports native login and Facebook login)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required:
                    - provider
                    - email
                    - password
                  properties:
                    provider:
                      type: string
                      enum: [native]
                      description: Login method
                    email:
                      type: string
                      format: email
                      description: User email
                    password:
                      type: string
                      description: User password
                - type: object
                  required:
                    - provider
                    - access_token
                  properties:
                    provider:
                      type: string
                      enum: [facebook]
                      description: Login method
                    access_token:
                      type: string
                      description: Facebook access token
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      access_token:
                        type: string
                      access_expired:
                        type: integer
                      user:
                        $ref: '#/components/schemas/User'
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                type: string
                example: "Email is not registered"
        "403":
          description: Authentication failed
          content:
            application/json:
              schema:
                type: string
                example: "Password is wrong"
  /api/v1/user/profile:
    get:
      tags:
        - Authentication
      summary: Get user profile
      description: Get current user profile information
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
                example: "no token"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
                example: "invalid token"

  # Product APIs
...
  /api/v1/report/payments:
    get:
      tags:
        - Reports
      summary: Get payments report
      description: Get payments analytics report
      responses:
        "200":
          description: Report retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    description: Payment report data
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: string
                example: "Failed to get payments report"
  /api/v1/report/total:
    get:
      tags:
        - Reports
      summary: Get total sales
      description: Get total sales amount
      responses:
        "200":
          description: Total sales retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: number
                    description: Total sales amount
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: string
                example: "Failed to get total sales"
  /api/v1/report/sold_color_percent:
    get:
      tags:
        - Reports
      summary: Get sold color percentage
      description: Get percentage of sold products by color
      responses:
        "200":
          description: Color percentage retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  soldColor:
                    type: object
                    description: Color percentage data
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: string
                example: "Failed to get sold color percent"
  /api/v1/report/sold_price_percent:
    get:
      tags:
        - Reports
      summary: Get sold price percentage
      description: Get percentage of sold products by price range
      responses:
        "200":
          description: Price percentage retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  soldPrice:
                    type: object
                    description: Price percentage data
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: string
                example: "Failed to get sold price percent"
  /api/v1/report/top-five:
    get:
      tags:
        - Reports
      summary: Get top five products
      description: Get top five selling products
      responses:
        "200":
          description: Top five products retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  topFive:
                    type: array
                    items:
                      type: object
                      description: Product sales data
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: string
                example: "Failed to get top five products"

  # Admin APIs
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        picture:
          type: string
        role_id:
          type: integer
...